% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_duplicates.R
\name{detect_duplicates}
\alias{detect_duplicates}
\title{Access 'observation' recordings and metadata}
\usage{
detect_duplicates(dataframe_1 = NULL, dataframe_2 = NULL, all_data = TRUE)
}
\arguments{
\item{dataframe_1}{dataframe refering to the metadata containing the multimedia information obtained from query functions.}

\item{dataframe_2}{dataframe refering to the metadata containing the multimedia information obtained from query functions.}

\item{all_data}{Logical argument that determines if all data available from database is shown in the results of search. Default is \code{TRUE}.}
}
\value{
If all_data is not provided the function returns a data frame with the following media
information: id, scientific_name, name, group, group_name, status, rarity, photo,
info_text, permalink, determination_requirements, file_url, repository
}
\description{
\code{detect_duplicates} detects duplicate data in data frames.
}
\details{
This function compares two dataframes to detect repeated data.
}
\examples{
\dontrun{
# compare
View(df3)
}

}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
