% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_dataset_update.R
\name{gbif_dataset_update}
\alias{gbif_dataset_update}
\alias{update_gbif_datasets}
\title{Access 'gbif' dataset information}
\usage{
gbif_dataset_update()
}
\value{
returns dataset csv from gbif
}
\description{
\code{gbif_dataset_update} searches for dataset information from \href{https://www.gbif.org}{gbif}.
}
\details{
This function queries for dataset info in the open-access
online repository \href{https://www.gbif.org}{gbif}.
}
\examples{
\dontrun{
# update dataset information
View(df1)

}

}
\references{
{

}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
