% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_metadata.R
\name{merge_metadata}
\alias{merge_metadata}
\title{Merge metadata data frames}
\usage{
merge_metadata(..., check_columns = TRUE)
}
\arguments{
\item{...}{two or more data frames (each one as a separate entry) referring
to the metadata obtained from suwo query functions (\code{query_x()}).
Alternatively, a single list of data frames can be provided. The name
provided for each data frame (either as individual data frames or in a list)
will be used as label in the \code{source} column in the output data frame.}

\item{check_columns}{Logical argument indicating if the function should
check that all input data frames have the required basic columns.
Default is \code{TRUE}.}
}
\value{
A single data frame with the data from all input data frames
combined and with an additional column named \code{source} indicating the
original data frame from which each row originated. The column \code{source} will
contain the name provided for each data frame (either as individual data
frames or in a list). If no names were provided, the object names will
be used instead.
}
\description{
\code{merge_metadata} merges metadata data frames from suwo queries.
}
\details{
This function combines metadata from multiple sources
(e.g. WikiAves and xeno-canto) into a single data frame for easier analysis
and comparison. Each input data frame must be obtained from one of the
suwo query functions (e.g., \code{query_wikiaves()}, \code{query_xenocanto()}, etc.)
with \code{raw_data = FALSE}.
}
\examples{
\donttest{
# get metadata from 2 repos
wa <- query_wikiaves(species = "Glaucis dohrnii", format =  "sound")
gb <- query_gbif(species = "Glaucis dohrnii", format = "sound")

# run if queries didnt fail
 if (!is.null(wa) && !is.null(gb)) {
 # combine metadata using single data frames
 merged_mt <- merge_metadata(wa, gb)

 # combine metadata using named single data frames
 merged_mt <- merge_metadata(wikiaves = wa, gbif = gb)

 # combine metadata using a list of data frames
 mt_list <- list(wikiaves = wa, gbif = gb)
 merged_mt <- merge_metadata(mt_list)
}
}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
