% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_query_results.R
\name{merge_query_results}
\alias{merge_query_results}
\title{Detect duplicate observations in suwo query results}
\usage{
merge_query_results(X, Y, all_data = getOption("all_data", FALSE))
}
\arguments{
\item{X}{dataframe referring to the metadata containing the multimedia information obtained from query functions.}

\item{Y}{a second dataframe referring to the metadata containing the multimedia information obtained from query functions.}

\item{all_data}{Logical argument that determines if all data available from database is shown in the results of search. Default is \code{FALSE}. Can be set globally for the current R session via the "all_data" option (see \code{\link[base]{options}}).}
}
\value{
If all_data is not provided the function returns a data frame with the following media
information: id, scientific_name, name, group, group_name, status, rarity, photo,
info_text, permalink, determination_requirements, file_url, repository
}
\description{
\code{merge_query_results} detects duplicate data in suwo query results.
}
\details{
This function compares two dataframes to detect repeated data.
}
\examples{
\dontrun{
# compare
View(df3)
}

}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
