% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_observation.R
\name{query_observation}
\alias{query_observation}
\title{Access 'observation' recordings and metadata}
\usage{
query_observation(term, type = c("sound", "still image", "moving image", "interactive resource"), cores = 1, pb = TRUE)
}
\arguments{
\item{term}{Character vector of length one indicating the
species, to query 'observation' database. For example \emph{Phaethornis longirostris}.
@param type Character vector with media type to query for. Currently 'still image' and 'sound' are available.}

\item{cores}{Numeric. Controls whether parallel computing is applied.
It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{pb}{Logical argument to control progress bar. Default is \code{TRUE}.}

\item{dataset}{see \url{https://observation.org/api/v1/species/search/?q=}}
}
\value{
If all_data is not provided the function returns a data frame with the following media information: id, scientific_name, name, group, group_name, status, rarity, photo, info_text, permalink, determination_requirements, file_url, repository
}
\description{
\code{query_observation} searches for metadata from \href{https://www.observation.org/}{observation}.
}
\details{
This function queries for species observation info in the open-access
online repository \href{https://www.observation.org/}{observation}. It can return media metadata.
}
\examples{
\dontrun{
# search without downloading
View(df1)

}

}
\references{
{

}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
