% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_metadata.R
\name{subset_metadata}
\alias{subset_metadata}
\title{Subset data frames while preserving attributes}
\usage{
subset_metadata(metadata, condition)
}
\arguments{
\item{metadata}{data frame previously obtained from any suwo query function (i.e. `query_reponame()`).}

\item{condition}{An expression indicating the rows to keep or the indices of the rows to keep.}
}
\value{
returns a data frame which is a subset of the input 'metadata'.
}
\description{
\code{subset_metadata} subsets metadata data frames while preserving attributes.
}
\details{
This function subsets a data frame while preserving its attributes. This is important for handling metadata data frames created by query functions in this package, which contain critical attributes that should be retained after subsetting.
}
\examples{
\dontrun{
# Download metadata
 pant_xc <- query_xenocanto(term = "Phaethornis anthophilus")

# Subset metadata to include only those from Panama
subset_metadata(pant_xc, country == "Panama")
}

}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
