% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suwo-package.R
\docType{package}
\name{suwo-package}
\alias{suwo}
\alias{suwo-package}
\title{suwo: A package to streamline bioacoustic analysis}
\description{
The suwo package aims to simplify the  retrieval of nature media
(mostly photos, audio files and videos) across  multiple online biodiversity
databases. The five major media repositories  accessed by this package
(GBIF, iNaturalist, Macaulay Library, WikiAves, and Xeno-Canto)
collectively host more than 250 million media files. Such media
are increasingly used in diverse fields, ranging from ecology and
evolutionary biology (e.g., trait evolution) to wildlife monitoring and
conservation (e.g., for training species detection models). The ability to
access and download large amounts of media files and their associated
metadata from a single interface thus provides a uniquely powerful
resource for facilitating research and conservation efforts.
}
\details{
The main features of the package are:
\itemize{
\item Obtaining media metadata from online repositories
\item Downloading associated media files
\item Updating data sets with new records
}

License: GPL (>= 2)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/suwo/}
  \item \url{https://github.com/ropensci/suwo/}
  \item Report bugs at \url{https://github.com/ropensci/suwo/issues/}
}

}
\author{
Marcelo Araya-Salas, Jorge Elizondo-Calvo & Alejandro Rico-Guevara

Maintainer: Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
\keyword{internal}
