% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_code_search.R
\name{taxon_code_search}
\alias{taxon_code_search}
\title{Access 'taxon codes'}
\usage{
taxon_code_search(species_name = NULL, taxon_code_df = NULL)
}
\arguments{
\item{term}{species name vector of length one indicating the species, to query 'clements database. For example \emph{Phaethornis longirostris}.}
}
\value{
Returns the taxon code for the given species name if match found in the clements taxon code data base
}
\description{
\code{taxon_code_search} searches for metadata from \href{https://www.birds.cornell.edu/clementschecklist/introduction/updateindex/october-2023/download/}{clements}.
}
\details{
This function queries for taxon code information from the clements taxon code data base \href{https://www.birds.cornell.edu/clementschecklist/introduction/updateindex/october-2023/download/}{clements}. It can return taxon code data. Clements data base must be downloaded and available in the work directory.
}
\examples{
\dontrun{
# search
}

}
\references{
{

}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
