% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_query.R
\name{update_query}
\alias{update_query}
\title{Access 'observation' recordings and metadata}
\usage{
update_query(dataframe_1 = NULL)
}
\arguments{
\item{dataframe_1}{dataframe refering to the metadata containing the multimedia information obtained from query functions.}

\item{all_data}{Logical argument that determines if all data available from database is shown in the results of search. Default is \code{TRUE}.}
}
\value{
If all_data is not provided the function returns a data frame with the following media
information: id, scientific_name, name, group, group_name, status, rarity, photo,
info_text, permalink, determination_requirements, file_url, repository
}
\description{
\code{update_query} detects duplicate data in data frames and updates new query results.
}
\details{
This function updates a previous query to add new information from the corresponding database of the original search
}
\examples{
\dontrun{
# compare
View(df3)
}

}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
