% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_media.R
\name{download_media}
\alias{download_media}
\title{Download media files from repositories}
\usage{
download_media(metadata)
}
\arguments{
\item{metadata}{Data frame with a 'file_url' column and any other column listed in the file.name argument. Only the media listed in the data frame
will be downloaded (\code{download} argument is automatically set to \code{TRUE}). This can be used to select
the recordings to be downloaded based on their attributes.}
}
\value{
media files
}
\description{
\code{download_media} downloads recordings and metadata from \href{https://www.xeno-canto.org/}{Xeno-Canto}, \href{https://www.wikiaves.com/}{wikiaves} or \href{https://www.gbif.org/}{gbif}.
}
\details{
File downloading process can be interrupted and resume later as long as the working directory is the same.
 Maps of recording coordinates can be produced using
\code{\link{map_results}}.
}
\examples{
\dontrun{
}

}
\references{
{
}
}
\seealso{
\code{\link{map_results}},
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
