% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_media.R
\name{download_media}
\alias{download_media}
\title{Download media files from repositories}
\usage{
download_media(
  metadata,
  path = ".",
  pb = getOption("pb", TRUE),
  verbose = getOption("verbose", TRUE),
  cores = getOption("mc.cores", 1),
  overwrite = FALSE,
  folder_by = NULL
)
}
\arguments{
\item{metadata}{data frame previously obtained from any suwo query
function (i.e. `query_reponame()`).}

\item{path}{Directory path where the output media files will be saved.
By default files are saved into the current working directory (\code{"."}).}

\item{pb}{Logical argument to control if progress bar is shown. Default
is \code{TRUE}. Can be set globally for the current R session via the "pb"
option ( \code{options(pb = TRUE)}).}

\item{verbose}{Logical argument that determines if text is shown in
console. Default is \code{TRUE}. Can be set globally for the current R
session via the "verbose" option ( \code{options(verbose = TRUE)}).}

\item{cores}{Numeric vector of length 1. Controls whether parallel computing
is applied by specifying the number of cores to be used. Default is 1
(i.e. no parallel computing). Can be set globally for the current R session
via the "mc.cores" option (e.g.  \code{options(mc.cores = 2)}). Note that
some repositories might not support parallel queries from the same IP
address as it might be identified as denial-of-service cyberattack.}

\item{overwrite}{Logical. If TRUE, existing files (in \code{"path"}) with
the same name will be overwritten. Default is FALSE.}

\item{folder_by}{Character string with the name of a character or factor
column in the metadata data frame. If supplied the function will use the
unique values in that column to create subfolders within \code{"path"} and
the files will be downloaded into the corresponding folder. By default no
subfolders are created and all files are saved in the path provided. Missing
values (NAs) are saved in a folder called
\code{paste0("unknown_", folder_by)}. Special characters that are not
allowed in folder names will be modified or removed. If any of the folder
names already exist in \code{"path"}, they will be used as is.}
}
\value{
Downloads media files into the supplied directory path
(\code{"path"}) and returns (invisibly) the input data frame with
two additional columns: \code{downloaded_file_name} with the name of
the downloaded file (if downloaded or already in the directory), and
\code{download_status} with the result of the download process for each
file (either "saved", "overwritten", "already there (not downloaded)",
or "failed").
}
\description{
\code{download_media} downloads recordings and metadata from
\href{https://www.xeno-canto.org/}{Xeno-Canto},
\href{https://www.wikiaves.com/}{wikiaves} or
\href{https://www.gbif.org/}{gbif}.
}
\details{
This function will take the output data frame of any of the
"query_x()" functions and download the associated media files. The
function will download all files into a single directory
(argument \code{"path"}). File downloading process can be interrupted and
resume later as long as the working directory is the same. By default only
the missing files will be downloaded when resuming. Users only need to rerun
the same function call. Can also be used on a updated query output
(see \code{\link{update_metadata}}) to add the new media files to the
existing media pool.
}
\examples{
if (interactive()){
  phae_anth <- query_xenocanto(species = 'Phaethornis anthophilus',
  all_data = FALSE)

  # donwload the first to files
  phae_anth_downl <- download_media(metadata = phae_anth[1:2, ],
  path = tempdir())
}

}
\seealso{
\code{\link{query_gbif}}, \code{\link{query_macaulay}}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
