% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_media.R
\name{download_media}
\alias{download_media}
\title{Download media files from repositories}
\usage{
download_media(
  metadata,
  path = ".",
  cores = getOption("suwo_cores", 1),
  pb = getOption("suwo_pb", TRUE),
  verbose = getOption("suwo_verbose", TRUE),
  overwrite = FALSE,
  folder_by = NULL
)
}
\arguments{
\item{metadata}{Data frame with the metadata of the media records to be
mapped. Typically the output of one of the query functions in this package
(e.g. \code{\link[=query_gbif]{query_gbif()}}, \code{\link[=query_inaturalist]{query_inaturalist()}}, etc.) or metadata formatting
functions (e.g. \code{\link[=merge_metadata]{merge_metadata()}}, \code{\link[=remove_duplicates]{remove_duplicates()}}, etc.).}

\item{path}{Directory path where the output media files will be saved.
By default files are saved into the current working directory (\code{"."}).}

\item{cores}{Numeric vector of length 1. Controls whether parallel computing
is applied by specifying the number of cores to be used. Default is 1
(i.e. no parallel computing). Can be set globally for the current R session
via the "mc.cores" option (e.g.  \code{options(mc.cores = 2)}). Note that
some repositories might not support parallel queries from the same IP
address as it might be identified as denial-of-service cyberattack.}

\item{pb}{Logical argument to control if progress bar is shown. Default
is \code{TRUE}. Can be set globally for the current R session via the "suwo_pb"
option ( \code{options(suwo_pb = TRUE)}). Not shown if only a few observations
are found.}

\item{verbose}{Logical argument that determines if text is shown in
console. Default is \code{TRUE}. Can be set globally for the current R
session via the "suwo_verbose" option
( \code{options(suwo_verbose = TRUE)}).}

\item{overwrite}{Logical. If TRUE, existing files (in \code{"path"}) with
the same name will be overwritten. Default is FALSE.}

\item{folder_by}{Character string with the name of a character or factor
column in the metadata data frame. If supplied the function will use the
unique values in that column to create subfolders within \code{"path"} and
the files will be downloaded into the corresponding folder. It can be used
for instance to save files in subfolders by country or recordist.
By default no  subfolders are created and all files are saved in the path
provided. Missing values (NAs) are saved in a folder called
\code{paste0("unknown_", folder_by)}. Special characters that are not
allowed in folder names will be modified or removed. If any of the folder
names already exist in \code{"path"}, they will be used as is.}
}
\value{
Downloads media files into the supplied directory path
(\code{"path"}) and returns (invisibly) the input data frame with
three additional columns: \code{downloaded_file_name} with the name of
the downloaded file (if downloaded or already in the directory),
\code{download_status} with the result of the download process for each
file (either "saved", "overwritten", "already there (not downloaded)",
or "failed"), and \code{file_size} with the size of the downloaded file in
megabytes (MB). Note that the column \code{file_size} can be used to further
remove duplicated media using \code{\link[=find_duplicates]{find_duplicates()}} and \code{\link[=remove_duplicates]{remove_duplicates()}}
(e.g. if two files from different repositories
have the same size and similar user name).
}
\description{
\code{download_media} downloads media files from online repositories.
}
\details{
This function will take the output data frame of any of the
"query_reponame()" functions and download the associated media files. The
function will download all files into a single directory
(argument \code{"path"}). File downloading process can be interrupted and
resume later as long as the working directory is the same. Users only need
to rerun the same function call. By default only the missing files will be
downloaded when resuming. Can also be used on a updated query output
(see \code{\link[=update_metadata]{update_metadata()}}) to add the new media files to the
existing media pool.
}
\examples{
 \donttest{
  a_zambiana <- query_inaturalist(species = "Amanita zambiana",
  format = "image")

# run if query didnt fail
 if (!is.null(a_zambiana)) {
  # donwload the first to files
  phae_anth_downl <- download_media(metadata = a_zambiana[1:2, ],
  path = tempdir())
}
}

}
\seealso{
\code{\link[=query_gbif]{query_gbif()}}, \code{\link[=query_macaulay]{query_macaulay()}}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
