% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_locations.R
\name{map_locations}
\alias{map_locations}
\title{Maps of instances of observations by species}
\usage{
map_locations(
  metadata,
  cluster = FALSE,
  palette = viridis::viridis,
  by = "species"
)
}
\arguments{
\item{metadata}{data frame previously obtained from any suwo query function (i.e. `query_reponame()`).}

\item{cluster}{Logical to control if icons are clustered by locality. Default is \code{FALSE}.}

\item{palette}{Color palette function used for location markers.}

\item{by}{Name of column to be used for coloring markers. Default is "species".}
}
\value{
An interacrive map with the locations of the observations.
}
\description{
\code{map_locations} creates maps to visualize the geographic spread of suwo
  recordings.
}
\details{
This function creates maps for visualizing the geographic spread of observations. Note that only recordings with geographic coordinates are displayed.
}
\examples{
\dontrun{
# search in xeno-canto
metadata <- query_xenocanto(term = "Phaethornis anthophilus")

# create map
map_locations(metadata)
}

}
\references{
{
Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.
}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr}) and Grace Smith Vidaurre
}
