% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_locations.R
\name{map_locations}
\alias{map_locations}
\title{Maps of media records}
\usage{
map_locations(
  metadata,
  cluster = FALSE,
  palette = grDevices::hcl.colors,
  by = "species"
)
}
\arguments{
\item{metadata}{data frame previously obtained from any suwo query
function (i.e. \code{query_reponame()}).}

\item{cluster}{Logical to control if icons are clustered by locality.
Default is \code{FALSE}.}

\item{palette}{Color palette function used for location markers. By default
it uses the virdis palette (\code{grDevices::hcl.colors}).}

\item{by}{Name of column to be used for coloring markers. Default is
"species".}
}
\value{
An interacrive map with the locations of the observations.
}
\description{
\code{map_locations} creates maps to visualize the geographic spread of
media records.
}
\details{
This function creates maps for visualizing the geographic
spread of observations. Note that only observations with geographic
coordinates are displayed.
}
\examples{
# search in xeno-canto
e_hochs <- query_gbif(species = "Entoloma hochstetteri", format = "image")

# run if query didnt fail
 if (!is.null(e_hochs)) {

# create map
map_locations(e_hochs)
}

}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr}) and Grace
Smith Vidaurre
}
