% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_metadata.R
\name{merge_metadata}
\alias{merge_metadata}
\title{Merge query results}
\usage{
merge_metadata(...)
}
\arguments{
\item{...}{two or more data frames (each one a a separate entry) referring to the metadata obtained from suwo query functions (`query_x()`).}
}
\value{
A single data frame with the data from all input data frames combined and with an additional column named `source` indicating the original data frame from which each row originated.
}
\description{
\code{merge_metadata} merges the output of suwo query results.
}
\details{
This function compares two dataframes to detect repeated data.
}
\examples{
\dontrun{
 # get metadata from 2 repos
  wa <- query_wikiaves(term = 'Glaucis dohrnii', format =  "sound")
  xc <- query_xenocanto(term = 'Glaucis dohrnii')

  # combine metadata
  merged_mt <- merge_metadata(wa, xc)
}

}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
