% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_metadata.R
\name{merge_metadata}
\alias{merge_metadata}
\title{Merge metadata data frames}
\usage{
merge_metadata(...)
}
\arguments{
\item{...}{two or more data frames (each one as a separate entry) referring
to the metadata obtained from suwo query functions (`query_x()`).}
}
\value{
A single data frame with the data from all input data frames
combined and with an additional column named `source` indicating the
original data frame from which each row originated.
}
\description{
\code{merge_metadata} merges metadata data frames from suwo queries.
}
\details{
This function combines metadata from multiple sources
(e.g. WikiAves and xeno-canto) into a single data frame for easier analysis
and comparison. Each input data frame must be obtained from one of the
suwo query functions (e.g., `query_wikiaves()`, `query_xenocanto()`, etc.)
with `raw_data = FALSE`.
}
\examples{
# get metadata from 2 repos
wa <- query_wikiaves(species = "Glaucis dohrnii", format =  "sound")
gb <- query_gbif(species = "Glaucis dohrnii", format = "sound")

# run if queries didnt fail
 if (!is.null(wa) && !is.null(gb)) {
 # combine metadata
 merged_mt <- merge_metadata(wa, gb)
}

}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
