% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_gbif.R
\name{query_gbif}
\alias{query_gbif}
\title{Access 'gbif' recordings and metadata}
\usage{
query_gbif(term = NULL, type = c("sound", "still image", "moving image", "interactive resource"), cores = 1, pb = TRUE, verbose = TRUE, dataset = NULL, all_data = TRUE)
}
\arguments{
\item{term}{Character vector of length one indicating genus and
species, to query 'gbif' database. For example, \emph{Phaethornis longirostris}.}

\item{type}{Character vector with media type to query for. Options are 'sound', 'stillimage', 'movingimage' and 'interactiveresource'. Required.}

\item{cores}{Numeric. Controls whether parallel computing is applied.
It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{pb}{Logical argument to control progress bar. Default is \code{TRUE}.}

\item{dataset}{see \url{https://www.gbif.org/dataset/search?q=}.}
}
\value{
If all_data is false the function returns a data frame with the following media information: id, species, date, country, location, latitude, longitude, file_url, repository
}
\description{
\code{query_gbif} searches for metadata from \href{https://www.gbif.org/}{gbif}.
}
\details{
This function queries for species observation info in the open-access
online repository \href{https://www.gbif.org/}{gbif}. It can return media metadata.
}
\examples{
\dontrun{
# search without downloading
View(df1)

}

}
\references{
{

}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
