% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_gbif.R
\name{query_gbif}
\alias{query_gbif}
\title{Access 'gbif' recordings and metadata}
\usage{
query_gbif(
  term = getOption("term", NULL),
  format = c("sound", "image", "video", "interactive resource"),
  cores = getOption("mc.cores", 1),
  pb = getOption("pb", TRUE),
  verbose = getOption("verbose", TRUE),
  dataset = NULL,
  all_data = getOption("all_data", FALSE),
  raw_data = getOption("raw_data", FALSE)
)
}
\arguments{
\item{term}{Character vector of length one indicating species, to query the database. For example, \emph{Phaethornis longirostris}. Required. Can be set globally for the current R session via the "term" option (e.g. \code{options(term = "Zonotrichia capensis")}).}

\item{format}{Character vector with the media format to query for. Options are 'sound', 'image', 'video' and 'interactive resource'. Required.}

\item{cores}{Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing). Can be set globally for the current R session via the "mc.cores" option (e.g.  \code{options(mc.cores = 2)}). Note that some repositories might not support parallel queries from the same IP address and errors might occur during the search.}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}. Can be set globally for the current R session via the "pb" option ( \code{options(pb = TRUE)}).}

\item{verbose}{Logical argument that determines if text is shown in console. Default is \code{TRUE}. Can be set globally for the current R session via the "verbose" option ( \code{options(verbose = TRUE)}).}

\item{dataset}{see \url{https://www.gbif.org/dataset/search?q=}.}

\item{all_data}{Logical argument that determines if all data available from database is shown in the results of search. Default is \code{FALSE}. Can be set globally for the current R session via the "all_data" option ( \code{options(all_data = TRUE)}).}

\item{raw_data}{Logical argument that determines if the raw data from the API is returned (e.g. without any manipulation). Default is \code{FALSE}. Can be set globally for the current R session via the "raw_data" option ( \code{options(raw_data = TRUE)}). If \code{TRUE}  \code{all_data} is set to \code{TRUE} internally. Useful for developers, or if users suspect that some data is mishandled during processing (i.e. date information is lost). Note that the metadata obtained when \code{raw_data = TRUE} is not standardized, so most suwo functions for downstream steps will not work on them.}
}
\value{
If all_data is \code{FALSE} the function returns a data frame with the metadata of media files matching the search term. If \code{all_data = TRUE}, all metadata fields are returned. If \code{raw_data = TRUE}, the raw data as obtained from the repository is returned (without any formatting).
}
\description{
\code{query_gbif} searches for metadata from \href{https://www.gbif.org/}{gbif}.
}
\details{
This function queries for species observation info in the open-access
online repository \href{https://www.gbif.org/}{gbif}. GBIF (the Global Biodiversity Information Facility) is an international network and data infrastructure funded by the world's governments and aimed at providing open access to data about all types of life on Earth. Note that some of the records returned by this function could be duplicates of records returned by other suwo functions (e.g., \code{\link{query_inaturalist}}).
}
\examples{
\dontrun{
# search sound files
View(df1)
}

}
\references{
{
GBIF.org (2024), GBIF Home Page. Available from: https://www.gbif.org [13 January 2020].
}
}
\seealso{
\code{\link{query_gbif}}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
