% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_gbif.R
\name{query_gbif}
\alias{query_gbif}
\title{Access 'gbif' recordings and metadata}
\usage{
query_gbif(
  term = NULL,
  type = c("sound", "still image", "moving image", "interactive resource"),
  cores = 1,
  pb = TRUE,
  verbose = TRUE,
  dataset = NULL,
  all_data = TRUE,
  save_path = paste0(term, ".csv")
)
}
\arguments{
\item{term}{Character vector of length one indicating genus and
species, to query 'gbif' database. For example, \emph{Phaethornis longirostris}.}

\item{type}{Character vector with media type to query for. Options are 'sound', 'stillimage', 'movingimage' and 'interactiveresource'. Required.}

\item{cores}{Numeric. Controls whether parallel computing is applied.
It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{pb}{Logical argument to control progress bar. Default is \code{TRUE}.}

\item{verbose}{Logical argument that determines if text is shown in console. Default is \code{TRUE}.}

\item{dataset}{see \url{https://www.gbif.org/dataset/search?q=}.}

\item{all_data}{Logical argument that determines if all data available from database is shown in the results of search. Default is \code{TRUE}.}
}
\value{
If all_data is false the function returns a data frame with the following media information: id, species, date, country, location, latitude, longitude, file_url, repository

If all_data is \code{TRUE} the function returns a data frame with the following media
information: key, datasetKey, publishingOrgKey, installationKey, hostingOrganizationKey,
publishingCountry, protocol, lastCrawled, lastParsed, crawlId, basisOfRecord, occurrenceStatus,
taxonKey, kingdomKey, phylumKey, classKey, orderKey, familyKey, genusKey, speciesKey,
acceptedTaxonKey, scientificName, acceptedScientificName, kingdom, phylum, order, family, genus,
species, genericName, specificEpithet, taxonRank, taxonomicStatus, iucnRedListCategory,
dateIdentified, decimalLatitude, decimalLongitude, coordinateUncertaintyInMeters, continent,
stateProvince, year, month, day, eventDate, modified, lastInterpreted, references, license,
isInCluster, datasetName, recordedBy, identifiedBy, geodeticDatum, class, countryCode, country,
rightsHolder, identifier, http...unknown.org.nick, verbatimEventDate, verbatimLocality,
collectionCode, gbifID, occurrenceID, taxonID, catalogNumber, institutionCode, eventTime,
occurrenceRemarks, http...unknown.org.captive, identificationID, identificationRemarks,
distanceFromCentroidInMeters, informationWithheld, sex, lifeStage, preparations,
nomenclaturalCode, dynamicProperties, locality, vernacularName, fieldNotes, verbatimElevation,
behavior, higherClassification, associatedTaxa, infraspecificEpithet, media-type, media-format,
media-references, media-created, media-creator, media-publisher, media-license,
media-rightsHolder, file_url, media-description, page, elevation, elevationAccuracy,
organismQuantity, organismQuantityType, georeferenceProtocol, verbatimSRS, county,
verbatimCoordinateSystem, type, collectionID, individualCount, samplingProtocol,
scientificNameID, georeferenceRemarks, language, georeferenceSources, media-title, repository.
}
\description{
\code{query_gbif} searches for metadata from \href{https://www.gbif.org/}{gbif}.
}
\details{
This function queries for species observation info in the open-access
online repository \href{https://www.gbif.org/}{gbif}. It can return media metadata.
}
\examples{
\dontrun{
# search without downloading
View(df1)
}

}
\references{
{
GBIF.org (2024), GBIF Home Page. Available from: https://www.gbif.org [13 January 2020].
}
}
\seealso{
\code{\link{query_gbif}}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
