% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_inaturalist.R
\name{query_inaturalist}
\alias{query_inaturalist}
\title{Access 'inaturalist' recordings and metadata}
\usage{
query_inaturalist(
  term = getOption("term", NULL),
  cores = getOption("mc.cores", 1),
  pb = getOption("pb", TRUE),
  verbose = getOption("verbose", TRUE),
  format = c("sound", "image"),
  identified = FALSE,
  verifiable = FALSE,
  all_data = getOption("all_data", FALSE),
  raw_data = getOption("raw_data", FALSE)
)
}
\arguments{
\item{term}{Character vector of length one indicating species, to query 'inaturalist' database. Can be set globally for the current R session via the "term" option (e.g. \code{options(term = "Phaethornis longirostris")}).}

\item{cores}{Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing). Can be set globally for the current R session via the "mc.cores" option (e.g.  \code{options(mc.cores = 2)}). Note that some repositories might not support parallel queries from the same IP address and errors might occur during the search.}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}. Can be set globally for the current R session via the "pb" option ( \code{options(pb = TRUE)}).}

\item{verbose}{Logical argument that determines if text is shown in console. Default is \code{TRUE}. Can be set globally for the current R session via the "verbose" option ( \code{options(verbose = TRUE)}).}

\item{format}{Character vector with the media format to query for. Currently 'image' and 'sound' are available.}

\item{identified}{Logical argument to define if search results are categorized as identified by inaturalist.}

\item{verifiable}{Logical argument to define if search results are categorized as verifiable by inaturalist.}

\item{all_data}{Logical argument that determines if all data available from database is shown in the results of search. Default is \code{TRUE}.}

\item{raw_data}{Logical argument that determines if the raw data from the API is returned (e.g. without any manipulation). Default is \code{FALSE}. Can be set globally for the current R session via the "raw_data" option ( \code{options(raw_data = TRUE)}). If \code{TRUE}  \code{all_data} is set to \code{TRUE} internally. Useful for developers, or if users suspect that some data is mishandled during processing (i.e. date information is lost). Note that the metadata obtained when \code{raw_data = TRUE} is not standardized, so most suwo functions for downstream steps will not work on them.}
}
\value{
If all_data is not provided the function returns a data frame with the following media
information: quality_grade, time_observed_at, taxon_geoprivacy, uuid, id, cached_votes_total,
identifications_most_agree, species_guess, identifications_most_disagree, positional_accuracy,
comments_count, site_id, created_time_zone, license_code, observed_time_zone,
public_positional_accuracy, oauth_application_id, created_at, description, time_zone_offset,
observed_on, observed_on_string, updated_at, captive, faves_count, num_identification_agreements,
map_scale, uri, community_taxon_id, owners_identification_from_vision, identifications_count,
obscured, num_identification_disagreements, geoprivacy, locality, spam, mappable,
identifications_some_agree, place_guess, file_url, attribution, page, repository
}
\description{
\code{query_inaturalist} searches for metadata from \href{https://www.inaturalist.org/}{inaturalist}.
}
\details{
This function queries for species observation info in the open-access
online repository \href{https://www.inaturalist.org/}{inaturalist}. It can return media metadata.
}
\examples{
\dontrun{
# search without downloading
View(df1)
}

}
\references{
{
iNaturalist. Available from https://www.inaturalist.org. Accessed [date]
}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
