% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_inaturalist.R
\name{query_inaturalist}
\alias{query_inaturalist}
\title{Access 'iNaturalist' media file metadata}
\usage{
query_inaturalist(
  species = getOption("suwo_species"),
  format = getOption("suwo_format", c("image", "sound")),
  cores = getOption("suwo_cores", 1),
  pb = getOption("suwo_pb", TRUE),
  verbose = getOption("suwo_verbose", TRUE),
  all_data = getOption("suwo_all_data", FALSE),
  raw_data = getOption("suwo_raw_data", FALSE),
  identified = FALSE,
  verifiable = FALSE
)
}
\arguments{
\item{species}{Character string with the scientific name of a species in
the format: "Genus epithet". Required. Can be set globally for the current
R session via the "suwo_species" option
(e.g. \code{options(suwo_species = "Hypsiboas rufitelus")}).}

\item{format}{Character vector with the media format to query for.
Currently 'image' and 'sound' are available. Can be set globally for
the current R session via the "suwo_format" option
(e.g. \code{options(suwo_format = "image")}). Required.}

\item{cores}{Numeric vector of length 1. Controls whether parallel computing
is applied by specifying the number of cores to be used. Default is 1
(i.e. no parallel computing). Can be set globally for the current R session
via the "mc.cores" option (e.g.  \code{options(mc.cores = 2)}). Note that
some repositories might not support parallel queries from the same IP
address as it might be identified as denial-of-service cyberattack.}

\item{pb}{Logical argument to control if progress bar is shown. Default
is \code{TRUE}. Can be set globally for the current R session via the "suwo_pb"
option ( \code{options(suwo_pb = TRUE)}). Not shown if only a few observations
are found.}

\item{verbose}{Logical argument that determines if text is shown in
console. Default is \code{TRUE}. Can be set globally for the current R
session via the "suwo_verbose" option
( \code{options(suwo_verbose = TRUE)}).}

\item{all_data}{Logical argument that determines if all data available
from database is shown in the results of search. Default is \code{TRUE}.}

\item{raw_data}{Logical argument that determines if the raw data from the
repository is returned (e.g. without any manipulation).
Default is \code{FALSE}. Can be set globally for the current R session
via the "suwo_raw_data" option ( \code{options(suwo_raw_data = TRUE)}). If \code{TRUE}
\code{all_data} is set to \code{TRUE} internally. Useful for developers,
or if users suspect that some data is mishandled during processing
(i.e. date information is lost). Note that the metadata obtained
when \code{raw_data = TRUE} is not standardized, so most suwo functions
for downstream steps will not work on them.}

\item{identified}{Logical argument to define if search results are
categorized as identified by inaturalist.}

\item{verifiable}{Logical argument to define if search results are
categorized as verifiable by inaturalist.}
}
\value{
The function returns a data frame with the metadata of the media
files matching the search criteria. If \code{all_data = TRUE}, all metadata
fields (columns) are returned. If \code{raw_data = TRUE}, the raw data as
obtained from the repository is returned (without any formatting).
}
\description{
\code{query_inaturalist} searches for metadata from
\href{https://www.inaturalist.org/}{iNaturalist}.
}
\details{
This function queries for species observation info in the
open-access
online repository \href{https://www.inaturalist.org/}{iNaturalist}.
iNaturalist is a free, crowdsourced online platform for nature enthusiasts
to document and identify plants, animals, fungi, and other organisms in
the wild. Note that Inaturalist observations do not include a 'country'
field.
}
\examples{
# search Bleeding Tooth mushroom images

}
\references{
iNaturalist. Available from https://www.inaturalist.org.
(Accessed on 10-02-2026)
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
