% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_macaulay.R
\name{query_macaulay}
\alias{query_macaulay}
\title{Searches for media files in the Macaulay Library}
\usage{
query_macaulay(
  term,
  format = c("sound", "image", "video"),
  verbose = getOption("verbose", TRUE),
  all_data = getOption("all_data", FALSE),
  path = ".",
  files = NULL,
  dates = NULL,
  taxon_code_info = ml_taxon_code
)
}
\arguments{
\item{term}{Character vector of length one indicating species, to query the database. For example, \emph{Phaethornis longirostris}. Required.}

\item{format}{Character vector with the media format to query for. Options are 'sound', 'image' of 'video'. Required.}

\item{verbose}{Logical argument that determines if text is shown in console. Default is \code{TRUE}. Can be set globally for the current R session via the "verbose" option (see \code{\link[base]{options}}).}

\item{all_data}{Logical argument that determines if all data available from database is shown in the results of search. Default is \code{FALSE}. Can be set globally for the current R session via the "all_data" option (see \code{\link[base]{options}}).}

\item{path}{Directory path where the .csv file will be saved. By default it is saved into the current working directory (\code{"."}).}

\item{files}{Optional character vector with the name(s) of the .csv file(s) to read. If not provided, the function will open a browser window to the search results page, where the user must download a .csv file with the metadata.}

\item{dates}{Optional numeric vector with years to split the search. If provided, the function will perform separate queries for each date range (between consecutive date values) and combine the results. Useful for queries that return large number of results (i.e. > 10000 results limit). For example, to search for the species between 2010 to 2020 and between 2021 to 2025 use \code{dates = c(2010, 2020, 2025)}. If years contain decimals searches will be split by months within years as well.}

\item{taxon_code_info}{Data frame containing the taxon code information. By default the function will use the internal data frame \code{"ml_taxon_code"} included as example data in the package. This object contains the data from the Clement list from october 2024 (downloaded from \url{https://www.birds.cornell.edu/clementschecklist/introduction/updateindex/october-2024/2024-citation-checklist-downloads/}). If new versions of the list become available it will be updated in new package versions. However, if users need to update it they can download the new list version, read it in R as a data frame and provide it to the function through this argument.}
}
\value{
If all_data is not provided the function returns a data frame with the following media
information: id, scientific_name, name, group, group_name, status, rarity, photo,
info_text, permalink, determination_requirements, file_url, repository
}
\description{
\code{query_macaulay} searches for metadata from \href{https://https://www.macaulaylibrary.org/}{macaulay}.
}
\details{
This function queries for species observation info in the \href{https://https://www.macaulaylibrary.org/}{Macaulay Library} online repository and returns the metadata of media files matching the query. The Macaulay Library is the world’s largest repository of digital media (audio, photo, and video) of wildlife, and their habitats. The archive hosts more than 77 million images, 3 million sound recordings, and 350k videos, from more than 80k contributors, and is integrated with eBird, the world’s largest biodiversity dataset. This is an interactive function which opens a browser window to the search results page, where the user must download a .csv file with the metadata. The function then reads the .csv file and returns a data frame with the metadata.

Here are some instructions for using this function properly:
\itemize{
   \item After the .csv file is saved, the user must confirm that into the R console.
   \item \emph{If the file is saved overwritting a pre-existing file (i.e. same file name) the function will not detect it}
   \item The query term must be a species name
   \item A maximum of 10000 records per query can be returned, but this can be bypassed by using the \code{dates} argument to split the search into smaller date ranges
   \item Users must log in to the Macaulay Library/eBird account in order to access large batches of observations
   }
}
\examples{
\dontrun{
# query sounds
tur_ili <- query_macaulay(term = "Turdus iliacus", format = "sound",
path = tempdir())

# test a query with more than 10000 results paging by date
cal_cos <- query_macaulay(term = "Calypte costae", type = "image",
path = tempdir(), dates = c(1976, 2020, 2022, 2024, 2025, 2026))

# this is how the internal function that splits the search by year intervals works
# it can split by entire year intervals
suwo:::.date_ranges(x = c(1976, 2020, 2022, 2024, 2025, 2026))

# or by year-month intervals if dates have decimals
# (note that it cannot split across years)
suwo:::.date_ranges(x = seq(2020, 2026, length.out = 10))

## update clement list (note that this is actually the same list used in the
# current 'suwo' version, just for the sake of the example)

# url to the clements list version october 2024
# (split so it is not truncaded by CRAN)
clements_url <- paste0(
"https://www.birds.cornell.edu/clementschecklist/wp-content/uploads/2024/10",
"/Clements-v2024-October-2024-rev.csv"
)

# read list from url
new_clements <- read.csv(clements_url)

# provide "updated" clements list to query_macaulay()
tur_ili2 <- query_macaulay(term = "Turdus iliacus", format = "sound", taxon_code_info = new_clements, path = tempdir())
}

}
\references{
{
Scholes III, Ph.D. E (2015). Macaulay Library Audio and Video Collection. Cornell Lab of Ornithology. Occurrence dataset https://doi.org/10.15468/ckcdpy accessed via GBIF.org on 2024-05-09.
Clements, J. F., P. C. Rasmussen, T. S. Schulenberg, M. J. Iliff, T. A. Fredericks, J. A. Gerbracht, D. Lepage, A. Spencer, S. M. Billerman, B. L. Sullivan, M. Smith, and C. L. Wood. 2024. The eBird/Clements checklist of Birds of the World: v2024.
}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
