% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_macaulay.R
\name{query_macaulay}
\alias{query_macaulay}
\title{Access 'observation' recordings and metadata}
\usage{
query_macaulay(
  term,
  type = c("audio", "photo", "video"),
  cores = getOption("mc.cores", 1),
  pb = getOption("pb", TRUE),
  verbose = getOption("verbose", TRUE),
  all_data = getOption("all_data", TRUE)
)
}
\arguments{
\item{term}{Character vector of length one indicating the
species, to query 'observation' database. For example \emph{Phaethornis longirostris}.}

\item{type}{Character vector with media type to query for. Currently 'photo' and 'audio' are available.}

\item{cores}{Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing). Can be set globally for the current R session via the "mc.cores" option (see \code{\link[base]{options}}).}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}. Can be set globally for the current R session via the "pb" option (see \code{\link[base]{options}}).}

\item{verbose}{Logical argument that determines if text is shown in console. Default is \code{TRUE}. Can be set globally for the current R session via the "verbose" option (see \code{\link[base]{options}}).}

\item{all_data}{Logical argument that determines if all data available from database is shown in the results of search. Default is \code{TRUE}. Can be set globally for the current R session via the "all_data" option (see \code{\link[base]{options}}).}
}
\value{
If all_data is not provided the function returns a data frame with the following media
information: id, scientific_name, name, group, group_name, status, rarity, photo,
info_text, permalink, determination_requirements, file_url, repository
}
\description{
\code{query_macaulay} searches for metadata from \href{https://https://www.macaulaylibrary.org/}{macaulay}.
}
\details{
This function queries for species observation info in the open-access
online repository \href{https://https://www.macaulaylibrary.org/}{macaulay}. It can return media metadata.
}
\examples{
\dontrun{
# search without downloading
View(df1)
}

}
\references{
{
Scholes III, Ph.D. E (2015). Macaulay Library Audio and Video Collection. Cornell Lab of Ornithology. Occurrence dataset https://doi.org/10.15468/ckcdpy accessed via GBIF.org on 2024-05-09.
}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
