% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_wikiaves.R
\name{query_wikiaves}
\alias{query_wikiaves}
\title{Access 'wikiaves' recordings and metadata}
\usage{
query_wikiaves(
  term = NULL,
  type = c("sound", "still image"),
  cores = 1,
  pb = TRUE,
  verbose = TRUE,
  all_data = TRUE,
  save_path = paste0(term, ".csv")
)
}
\arguments{
\item{term}{Character vector of length one indicating species, to query 'wikiaves' database. For example, \emph{Phaethornis longirostris}.}

\item{type}{Character vector with media type to query for. Options are 'still image' or 'sound'. Required.}

\item{cores}{Numeric. Controls whether parallel computing is applied.
It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{pb}{Logical argument to control progress bar. Default is \code{TRUE}.}

\item{verbose}{Logical argument that determines if text is shown in console. Default is \code{TRUE}.}

\item{all_data}{Logical argument that determines if all data available from database is shown in the results of search. Default is \code{TRUE}.}
}
\value{
If all_data is not provided the function returns a data frame with the following
recording information: recording ID, media type, user ID, species ID, scientific name, common
name, repository ID, author, user name, date, verified condition, location, location ID,
comments, likes, visualizations, url, duration and repository
}
\description{
\code{query_wikiaves} searches for metadata from \href{https://www.wikiaves.com/}{wikiaves}.
}
\details{
This function queries for avian vocalization recordings in the open-access
online repository \href{https://www.wikiaves.com/}{wikiaves}. It can return recordings metadata.
}
\examples{
\dontrun{
# search
df1 <- query_wikiaves(term = "Phaethornis nattereri", type = "still image")
View(df1)
}

}
\references{
{
Schubert, Stephanie Caroline, Lilian Tonelli Manica, and André De Camargo Guaraldo. 2019. Revealing the potential of a huge citizen-science platform to study bird migration. Emu-Austral Ornithology 119.4: 364-373.
}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
