% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_wikiaves.R
\name{query_wikiaves}
\alias{query_wikiaves}
\title{Access 'Wikiaves' recordings and metadata}
\usage{
query_wikiaves(
  term = getOption("term", NULL),
  format = c("sound", "image"),
  cores = getOption("mc.cores", 1),
  pb = getOption("pb", TRUE),
  verbose = getOption("verbose", TRUE),
  all_data = getOption("all_data", FALSE),
  raw_data = getOption("raw_data", FALSE)
)
}
\arguments{
\item{term}{Character vector of length one indicating species, to query the database. For example, \emph{Phaethornis longirostris}. Required. Can be set globally for the current R session via the "term" option (e.g. \code{options(term = "Zonotrichia capensis")}).}

\item{format}{Character vector with the media format to query for. Options are 'sound' or 'image'. Required.}

\item{cores}{Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing). Can be set globally for the current R session via the "mc.cores" option (e.g.  \code{options(mc.cores = 2)}). Note that some repositories might not support parallel queries from the same IP address and errors might occur during the search.}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}. Can be set globally for the current R session via the "pb" option ( \code{options(pb = TRUE)}).}

\item{verbose}{Logical argument that determines if text is shown in console. Default is \code{TRUE}. Can be set globally for the current R session via the "verbose" option ( \code{options(verbose = TRUE)}).}

\item{all_data}{Logical argument that determines if all data available from database is shown in the results of search. Default is \code{FALSE}. Can be set globally for the current R session via the "all_data" option ( \code{options(all_data = TRUE)}).}

\item{raw_data}{Logical argument that determines if the raw data from the API is returned (e.g. without any manipulation). Default is \code{FALSE}. Can be set globally for the current R session via the "raw_data" option ( \code{options(raw_data = TRUE)}). If \code{TRUE}  \code{all_data} is set to \code{TRUE} internally. Useful for developers, or if users suspect that some data is mishandled during processing (i.e. date information is lost). Note that the metadata obtained when \code{raw_data = TRUE} is not standardized, so most suwo functions for downstream steps will not work on them.}
}
\value{
A data frame with the metadata of the observations matching the query.
}
\description{
\code{query_wikiaves} searches for metadata from \href{https://www.wikiaves.com/}{wikiaves}.
}
\details{
This function queries for avian digital media in the open-access
online repository \href{https://www.wikiaves.com/}{wikiaves} and returns its metadata. WikiAves is a Brazilian online platform and citizen science project that serves as the largest community for birdwatchers in Brazil. It functions as a collaborative, interactive encyclopedia of Brazilian birds, where users contribute georeferenced photographs and sound recordings, which are then used to build a vast database for research and conservation.
}
\examples{
\dontrun{
# search
df1 <- query_wikiaves(term = "Phaethornis nattereri", format = "image")
View(df1)
}

}
\references{
{
Schubert, Stephanie Caroline, Lilian Tonelli Manica, and André De Camargo Guaraldo. 2019. Revealing the potential of a huge citizen-science platform to study bird migration. Emu-Austral Ornithology 119.4: 364-373.
}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
