% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_xenocanto.R
\name{query_xenocanto}
\alias{query_xenocanto}
\title{Access 'Xeno-Canto' recordings and metadata}
\usage{
query_xenocanto(
  term = getOption("term", NULL),
  cores = getOption("mc.cores", 1),
  pb = getOption("pb", TRUE),
  verbose = getOption("verbose", TRUE),
  all_data = getOption("all_data", FALSE),
  raw_data = getOption("raw_data", FALSE)
)
}
\arguments{
\item{term}{Character vector of length one indicating the scientific of the taxonomic group (species, genus, or family)
to query for in the 'Xeno-Canto' database. For example, \emph{Phaethornis} or \emph{Phaethornis longirostris}. Can be set globally for the current R session via the "term" option (e.g. \code{options(term = "Phaethornis longirostris")}).
More complex queries can be done by using search terms that follow the
xeno-canto advance query syntax (only valid for this function). This syntax uses tags to search within a particular aspect of the recordings
(e.g. country, location, sound type). Tags are of the form tag:searchterm'. For instance, 'type:song'
will search for all recordings in which the sound type description contains the word 'song'.
Several tags can be included in the same query. The query "phaethornis cnt:belize' will only return
results for birds in the genus \emph{Phaethornis} that were recorded in Belize. Queries are case insensitive. Make sure taxonomy related tags (Genus or scientific name) are found first in multi-tag queries. See \href{https://www.xeno-canto.org/help/search}{Xeno-Canto's search help} for a full description and see examples below
for queries using terms with more than one word.}

\item{cores}{Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing). Can be set globally for the current R session via the "mc.cores" option (e.g.  \code{options(mc.cores = 2)}). Note that some repositories might not support parallel queries from the same IP address and errors might occur during the search.}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}. Can be set globally for the current R session via the "pb" option ( \code{options(pb = TRUE)}).}

\item{verbose}{Logical argument that determines if text is shown in console. Default is \code{TRUE}. Can be set globally for the current R session via the "verbose" option ( \code{options(verbose = TRUE)}).}

\item{all_data}{Logical argument that determines if all data available from database is shown in the results of search. Default is \code{FALSE}. Can be set globally for the current R session via the "all_data" option ( \code{options(all_data = TRUE)}).}

\item{raw_data}{Logical argument that determines if the raw data from the API is returned (e.g. without any manipulation). Default is \code{FALSE}. Can be set globally for the current R session via the "raw_data" option ( \code{options(raw_data = TRUE)}). If \code{TRUE}  \code{all_data} is set to \code{TRUE} internally. Useful for developers, or if users suspect that some data is mishandled during processing (i.e. date information is lost). Note that the metadata obtained when \code{raw_data = TRUE} is not standardized, so most suwo functions for downstream steps will not work on them.}
}
\value{
The function returns a data frame with the metadata of the matching observations. A description of the available fields from  Xeno-Canto can be found \href{https://xeno-canto.org/explore/api}{here}.
}
\description{
\code{query_xenocanto} searches for metadata from \href{https://www.xeno-canto.org/}{Xeno-Canto}.
}
\details{
This function queries for animal vocalization recordings in the open-access
online repository \href{https://www.xeno-canto.org/}{Xeno-Canto}. \href{https://www.xeno-canto.org/}{Xeno-Canto} is an online database that provides access to sound recordings of wildlife from around the world. The recordings are shared by a growing community of thousands of recordists from around the world.

 Complex queries can be done by using search terms that follow the
  \href{https://www.xeno-canto.org/}{Xeno-Canto} advance query syntax (check "term" argument description).
}
\examples{
\dontrun{
# search without downloading

## search using xeno-canto advance query ###
orth.pap <- query_xenocanto(term = "gen:orthonyx cnt:papua loc:tari")

# use quotes for queries with more than 1 word (e.g. Costa Rica),note that the
# single quotes are used for the whole 'term' and double quotes for the 2-word term inside
# Phaeochroa genus in Costa Rica
phae.cr <- query_xenocanto(term = 'gen:phaeochroa cnt:"costa rica"')

# several terms can be searched for in the same field
# search for all female songs in sound type
femsong <- query_xenocanto(term = "type:song type:female")
}

}
\references{
{
Planqué, Bob, & Willem-Pier Vellinga. 2008. Xeno-canto: a 21st-century way to appreciate Neotropical bird song. Neotrop. Birding 3: 17-23.
}
}
\seealso{
\code{\link{query_gbif}}, \code{\link{query_wikiaves}}, \code{\link{query_inaturalist}}, \code{\link{query_observation}}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
