% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_xenocanto.R
\name{query_xenocanto}
\alias{query_xenocanto}
\title{Access 'Xeno-Canto' recordings and metadata}
\usage{
query_xenocanto(term = NULL, cores = 1, pb = TRUE, verbose = TRUE, all_data = TRUE)
}
\arguments{
\item{term}{Character vector of length one indicating the scientific of the taxonomic group (species, genus, or family)
to query for in the 'Xeno-Canto' database. For example, \emph{Phaethornis} or \emph{Phaethornis longirostris}.
More complex queries can be done by using search terms that follow the
xeno-canto advance query syntax. This syntax uses tags to search within a particular aspect of the recordings
(e.g. country, location, sound type). Tags are of the form tag:searchterm'. For instance, 'type:song'
will search for all recordings in which the sound type description contains the word 'song'.
Several tags can be included in the same query. The query "phaethornis cnt:belize' will only return
results for birds in the genus \emph{Phaethornis} that were recorded in Belize. Queries are case insensitive. Make sure taxonomy related tags (Genus or scientific name) are found first in multi-tag queries. See \href{https://www.xeno-canto.org/help/search}{Xeno-Canto's search help} for a full description and see examples below
for queries using terms with more than one word.}

\item{cores}{Numeric. Controls whether parallel computing is applied.
It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{pb}{Logical argument to control progress bar. Default is \code{TRUE}.}

\item{verbose}{Logical argument that determines if text is shown in console. Default is \code{TRUE}.}

\item{all_data}{All}
}
\value{
The function returns a data frame with the following recording information: recording ID,
Genus, Specific epithet, Subspecies, English name, Recordist, Country, Locality, Latitude,
Longitude, Vocalization type, Audio file, License, URL, Quality, Time, Date. Sound files in .mp3
format are downloaded into the working directory if download = \code{TRUE}.
}
\description{
\code{query_xenocanto} searches for metadata from \href{https://www.xeno-canto.org/}{Xeno-Canto}.
}
\details{
This function queries for avian vocalization recordings in the open-access
online repository \href{https://www.xeno-canto.org/}{Xeno-Canto}. It can return recordings metadata
or download the associated sound files. Complex queries can be done by using search terms that follow the
 xeno-canto advance query syntax (check "term" argument description).
 Files are double-checked after downloading and "empty" files are re-downloaded.
 File downloading process can be interrupted and resume later as long as the working directory is the same.
 Maps of recording coordinates can be produced using
\code{\link{map_xc}}.
}
\examples{
\dontrun{
# search without downloading
df1 <- query_xenocanto(term = 'Phaethornis anthophilus')
## search using xeno-canto advance query ###
orth.pap <- query_xenocanto(term = 'gen:orthonyx cnt:papua loc:tari')


# use quotes for queries with more than 1 word (e.g. Costa Rica),note that the
# single quotes are used for the whole 'term' and double quotes for the 2-word term inside
#Phaeochroa genus in Costa Rica
phae.cr <- query_xenocanto(term = 'gen:phaeochroa cnt:"costa rica"')

# several terms can be searched for in the same field
# search for all female songs in sound type
femsong <- query_xenocanto(term = 'type:song type:female')
}

}
\references{
{
Planqué, Bob, & Willem-Pier Vellinga. 2008. Xeno-canto: a 21st-century way to appreciate Neotropical bird song. Neotrop. Birding 3: 17-23.
}
}
\seealso{
\code{\link{map_xc}},
\href{https://marce10.github.io/2016/12/22/Download_a_single_recording_for_each_species_in_a_site_from_Xeno-Canto.html}{blog post on accessing Xeno-Canto recordings}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
