% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_xenocanto.R
\name{query_xenocanto}
\alias{query_xenocanto}
\title{Access 'Xeno-Canto' recordings and metadata}
\usage{
query_xenocanto(
  term,
  other_tags = NULL,
  key,
  cores = getOption("mc.cores", 1),
  pb = getOption("pb", TRUE),
  verbose = getOption("verbose", TRUE),
  all_data = getOption("all_data", FALSE),
  raw_data = getOption("raw_data", FALSE)
)
}
\arguments{
\item{term}{Character vector of length one indicating the scientific species name
(e.g., "Phaethornis longirostris"). The function will automatically format this as 'sp:Phaethornis longirostris'
in the query.}

\item{other_tags}{Optional. A character vector containing additional tags to refine the search,
following the Xeno-Canto advanced query syntax. Tags are of the form 'tag:searchterm'.
For instance, 'type:song' will search for recordings where the sound type contains 'song'.
Multiple tags can be provided (e.g., \code{c("cnt:belize", "type:song")}).
See \href{https://www.xeno-canto.org/help/search}{Xeno-Canto's search help} for a full description.}

\item{cores}{Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing). Can be set globally for the current R session via the "mc.cores" option (e.g.  \code{options(mc.cores = 2)}). Note that some repositories might not support parallel queries from the same IP address and errors might occur during the search.}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}. Can be set globally for the current R session via the "pb" option ( \code{options(pb = TRUE)}).}

\item{verbose}{Logical argument that determines if text is shown in console. Default is \code{TRUE}. Can be set globally for the current R session via the "verbose" option ( \code{options(verbose = TRUE)}).}

\item{all_data}{Logical argument that determines if all data available from database is shown in the results of search. Default is \code{FALSE}. Can be set globally for the current R session via the "all_data" option ( \code{options(all_data = TRUE)}).}

\item{raw_data}{Logical argument that determines if the raw data from the API is returned (e.g. without any manipulation). Default is \code{FALSE}. Can be set globally for the current R session via the "raw_data" option ( \code{options(raw_data = TRUE)}). If \code{TRUE}  \code{all_data} is set to \code{TRUE} internally. Useful for developers, or if users suspect that some data is mishandled during processing (i.e. date information is lost). Note that the metadata obtained when \code{raw_data = TRUE} is not standardized, so most suwo functions for downstream steps will not work on them.}
}
\value{
The function returns a data frame with the following recording information: recording ID,
 Genus, Specific epithet, Subspecies, English name, Recordist, Country, Locality, Latitude,
 Longitude, Vocalization type, Audio file, License, URL, Quality, Time, and Date.
}
\description{
\code{query_xenocanto} searches for metadata from \href{https://www.xeno-canto.org/}{Xeno-Canto}.
}
\details{
This function queries for avian vocalization recordings in the open-access
 online repository \href{https://www.xeno-canto.org/}{Xeno-Canto}. It can return recordings metadata
 or download the associated sound files. Complex queries can be constructed by combining the
 \code{term} and \code{other_tags} arguments.
}
\examples{
\dontrun{
# An API key is required. Get yours at https://xeno-canto.org/account.
XC_API_KEY <- "YOUR_API_KEY_HERE"

# Simple search for a species (will be converted to sp:"Phaethornis anthophilus")
df1 <- query_xenocanto(term = "Phaethornis anthophilus", key = XC_API_KEY)

# Search for a term and add other tags for country and quality grade
pany.cr <- query_xenocanto(term = "Panyptila cayennensis",
                           other_tags = c('cnt:"costa rica"', "q:A"),
                           key = XC_API_KEY)

# Search for female songs of a species
femsong <- query_xenocanto(term = "Thryothorus ludovicianus",
                           other_tags = c("type:song", "type:female"),
                           key = XC_API_KEY)
}

}
\references{
{
Planqué, Bob, & Willem-Pier Vellinga. 2008. Xeno-canto: a 21st-century way to appreciate Neotropical bird song. Neotrop. Birding 3: 17-23.
}
}
\seealso{
\code{\link{query_gbif}}, \code{\link{query_wikiaves}}, \code{\link{query_inaturalist}}, \code{\link{query_observation}}
\href{https://marce10.github.io/2016/12/22/Download_a_single_recording_for_each_species_in_a_site_from_Xeno-Canto.html}{blog post on accessing Xeno-Canto recordings}
}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
