% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_metadata.R
\name{subset_metadata}
\alias{subset_metadata}
\title{Subset data frames while preserving attributes}
\usage{
subset_metadata(metadata, subset, select)
}
\arguments{
\item{metadata}{data frame previously obtained from any suwo query function (i.e. `query_reponame()`).}

\item{subset}{logical expression or numeric vector (with row numbers) indicating rows to keep.}

\item{select}{logical expression, numeric vector (with column numbers) or character vector (with column names) indicating columns to keep.}
}
\value{
returns a data frame which is a subset of the input 'metadata'.
}
\description{
\code{subset_metadata} subsets metadata data frames while preserving attributes.
}
\details{
This function subsets a data frame while preserving its attributes. This is important for handling metadata data frames created by query functions in this package, which contain critical attributes that should be retained after subsetting.
}
\examples{
\dontrun{
# Download metadata
 pant_xc <- query_xenocanto(term = "Phaethornis anthophilus")

# Subset metadata to include only those from Panama
subset_metadata(pant_xc, subset = country == "Panama")

# get first two rows
subset_metadata(pant_xc, subset = 1:2)

# select only specific columns
subset_metadata(pant_xc, select = c("species", "date", "country", "file_url"))
}

}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
