% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suwo-package.R
\docType{package}
\name{suwo-package}
\alias{suwo}
\alias{suwo-package}
\title{suwo: A package to streamline bioacoustic analysis}
\description{
suwo is intended to facilitate the search of biological observation data
across multiple databases
}
\details{
The main features of the package are:
  \itemize{
  \item The use of loops to apply tasks through acoustic signals referenced
  in a selection table
  \item The production of images in the working folder with spectrograms
  that allow to organize data and verify acoustic analyzes
  }

The package offers functions to:
  \itemize{
  \item Explore and download observational data from iNaturalist, GBIF,
  Macaulay Library, Observation.org, Wikiaves and Xenocanto.
  \item Download different types of data across multiple databases
  \item Consolidate databases and annotation tables
  }

Most of the functions allow the parallelization of tasks, which distributes
the tasks among several processors to improve computational efficiency.
Tools to evaluate the performance of the analysis at each step are also
available. In addition, warbleR satisfies the need for rigorous open source
bioacoustic analysis, which facilitates opportunities for use in research
and innovation of additional custom analyzes.

License: GPL (>= 2)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://marce10.github.io/suwo/}
  \item \url{https://github.com/maRce10/suwo}
  \item Report bugs at \url{https://github.com/maRce10/suwo/issues/}
}

}
\author{
Marcelo Araya-Salas, Jorge Elizondo-Calvo & Alejandro Rico-Guevara

  Maintainer: Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
\keyword{internal}
