% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_query.R
\name{update_query}
\alias{update_query}
\title{Access 'observation' recordings and metadata}
\usage{
update_query(previous_query, token)
}
\arguments{
\item{previous_query}{data frame referring to the metadata containing the multimedia information previously obtained from query functions.}

\item{token}{Character referring to the token assigned by Observation.org as authorization for searches.}
}
\description{
\code{update_query} detects duplicate data in data frames and updates new query results.
}
\details{
This function updates a previous query to add new information from the corresponding database of the original search
}
\examples{
\dontrun{
# compare
View(df3)
}

}
\author{
Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
